+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+**Battlegrounds**:
+- Fixed empty Lockbox filter appearing in non-Pirate lobbies.
+
 ## **Release v1.55.13 - 2026-08-17**
 **General**:
 - Fixed not all UI elements updating immediately when you changed the deck tracker's language.